feat: pool updates
