+++ .env
@@ -1,4 +1,4 @@
-REACT_APP_NETWORK_URL="https://bsc-dataseed1.ninicoin.io/"
+REACT_APP_NETWORK_URL="https://bsc.mytokenpocket.vip"
 REACT_APP_CHAIN_ID="56"
 REACT_APP_GTAG = "GTM-TLF66T4"
 SKIP_PREFLIGHT_CHECK=true
