+++ .env.production
@@ -1,14 +1,24 @@
-REACT_APP_CHAIN_ID = "56"
-REACT_APP_GTAG = "GTM-TLF66T4"
+NEXT_PUBLIC_CHAIN_ID = "56"
+NEXT_PUBLIC_GTAG = "GTM-TLF66T4"
 
 # 10+ nodes balanced, US/EU
-REACT_APP_NODE_1 = "https://bsc-dataseed1.ninicoin.io"
+NEXT_PUBLIC_NODE_1 = "https://bsc-dataseed1.ninicoin.io"
 
 # 10+ nodes balanced, US/EU
-REACT_APP_NODE_2 = "https://bsc-dataseed1.defibit.io"
+NEXT_PUBLIC_NODE_2 = "https://bsc-dataseed1.defibit.io"
 
 # 10+ nodes balanced in each region, global
-REACT_APP_NODE_3 = "https://bsc-dataseed.binance.org"
+NEXT_PUBLIC_NODE_3 = "https://bsc-dataseed.binance.org"
 
-REACT_APP_GRAPH_API_PROFILE = "https://api.thegraph.com/subgraphs/name/pancakeswap/profile"
-REACT_APP_GRAPH_API_PREDICTION = "https://api.thegraph.com/subgraphs/name/pancakeswap/prediction"
+# Google Cloud Infrastructure Endpoint - Global
+NEXT_PUBLIC_NODE_PRODUCTION = "https://nodes.pancakeswap.com"
+
+NEXT_PUBLIC_GRAPH_API_PROFILE = "https://api.thegraph.com/subgraphs/name/pancakeswap/profile"
+NEXT_PUBLIC_GRAPH_API_PREDICTION = "https://api.thegraph.com/subgraphs/name/pancakeswap/prediction-v2"
+NEXT_PUBLIC_GRAPH_API_LOTTERY = "https://api.thegraph.com/subgraphs/name/pancakeswap/lottery"
+NEXT_PUBLIC_GRAPH_API_NFT_MARKET = "https://api.thegraph.com/subgraphs/name/pancakeswap/nft-market"
+
+NEXT_PUBLIC_SNAPSHOT_BASE_URL = "https://hub.snapshot.org"
+
+NEXT_PUBLIC_API_NFT = "https://nft.pancakeswap.com/api/v1"
+NEXT_PUBLIC_BIT_QUERY_ENDPOINT = "https://graphql.bitquery.io"
