+++ src/views/Pools/components/PoolCard/CardFooter/index.tsx
@@ -21,6 +21,7 @@ interface FooterProps {
   account: string
   isAutoVault?: boolean
   totalCakeInVault?: BigNumber
+  lockTime?: any
 }
 
 const ExpandableButtonWrapper = styled(Flex)`
@@ -31,7 +32,7 @@ const ExpandableButtonWrapper = styled(Flex)`
   }
 `
 
-const Footer: React.FC<FooterProps> = ({ pool, account, isAutoVault = false }) => {
+const Footer: React.FC<FooterProps> = ({ pool, account, isAutoVault = false, lockTime }) => {
   const { t } = useTranslation()
   const [isExpanded, setIsExpanded] = useState(false)
 
@@ -79,7 +80,7 @@ const Footer: React.FC<FooterProps> = ({ pool, account, isAutoVault = false }) =
           {isExpanded ? <Text>Hide</Text> : <Text>Details</Text>}
         </ExpandableLabel>
       </ExpandableButtonWrapper>
-      {isExpanded && <ExpandedFooter pool={pool} account={account} isAutoVault={isAutoVault} />}
+      {isExpanded && <ExpandedFooter pool={pool} account={account} isAutoVault={isAutoVault} lockTime={lockTime} />}
     </CardFooter>
   )
 }
