+++ vercel.json
@@ -5,6 +5,10 @@
       "destination": "/$1"
     },
     {
+      "source": "/bsc/staking/(.*)",
+      "destination": "/$1"
+    },
+    {
       "source": "/bsc/(.*)",
       "destination": "/$1"
     }
@@ -28,7 +32,7 @@
         { "key": "Access-Control-Allow-Credentials", "value": "true" },
         {
           "key": "Access-Control-Allow-Origin",
-          "value": "https://bsc-dataseed1.ninicoin.io/"
+          "value": "*"
         },
         {
           "key": "Access-Control-Allow-Methods",
@@ -36,7 +40,7 @@
         },
         {
           "key": "Access-Control-Allow-Headers",
-          "value": "X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version"
+          "value": "*"
         }
       ]
     },
